refactor(http)!: reduce api integer sizes (#1505)
Validation constants that refer to amounts of seconds or amounts of
users did not necessarily need to take an entire `u64`, when their
values never surpass 50,000. This PR reduces them to `u16` in `http` and
`validate`, while leaving `model` alone.

+++ http/src/request/channel/invite/create_invite.rs
@@ -21,9 +21,9 @@ use twilight_validate::request::{
 #[derive(Serialize)]
 struct CreateInviteFields {
     #[serde(skip_serializing_if = "Option::is_none")]
-    max_age: Option<u64>,
+    max_age: Option<u32>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    max_uses: Option<u64>,
+    max_uses: Option<u16>,
     #[serde(skip_serializing_if = "Option::is_none")]
     temporary: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -114,7 +114,7 @@ impl<'a> CreateInvite<'a> {
     /// println!("invite code: {}", invite.code);
     /// # Ok(()) }
     /// ```
-    pub const fn max_age(mut self, max_age: u64) -> Result<Self, ValidationError> {
+    pub const fn max_age(mut self, max_age: u32) -> Result<Self, ValidationError> {
         if let Err(source) = validate_invite_max_age(max_age) {
             return Err(source);
         }
@@ -150,7 +150,7 @@ impl<'a> CreateInvite<'a> {
     /// println!("invite code: {}", invite.code);
     /// # Ok(()) }
     /// ```
-    pub const fn max_uses(mut self, max_uses: u64) -> Result<Self, ValidationError> {
+    pub const fn max_uses(mut self, max_uses: u16) -> Result<Self, ValidationError> {
         if let Err(source) = validate_invite_max_uses(max_uses) {
             return Err(source);
         }
