+++ http/src/routing.rs
@@ -29,7 +29,7 @@ pub enum Route<'a> {
     CreateBan {
         /// The number of days' worth of the user's messages to delete in the
         /// guild's channels.
-        delete_message_days: Option<u64>,
+        delete_message_days: Option<u16>,
         /// The ID of the guild.
         guild_id: u64,
         /// The ID of the user.
@@ -75,7 +75,7 @@ pub enum Route<'a> {
         compute_prune_count: Option<bool>,
         /// The number of days that a user must be offline before being able to
         /// be pruned.
-        days: Option<u64>,
+        days: Option<u16>,
         /// The ID of the guild.
         guild_id: u64,
         /// The roles to filter the prune by.
@@ -350,7 +350,7 @@ pub enum Route<'a> {
         /// The ID of the guild.
         guild_id: u64,
         /// The maximum number of audit logs to get.
-        limit: Option<u64>,
+        limit: Option<u16>,
         /// The ID of the user, if specified.
         user_id: Option<u64>,
     },
@@ -496,7 +496,7 @@ pub enum Route<'a> {
         /// The ID of the guild.
         guild_id: u64,
         /// The maximum number of members to get.
-        limit: Option<u64>,
+        limit: Option<u16>,
         /// Whether to get the members' presences.
         presences: Option<bool>,
     },
@@ -510,7 +510,7 @@ pub enum Route<'a> {
     GetGuildPruneCount {
         /// The number of days that a user must be offline before being able to
         /// be pruned.
-        days: Option<u64>,
+        days: Option<u16>,
         /// The ID of the guild.
         guild_id: u64,
         /// The roles to filter the prune by.
@@ -549,7 +549,7 @@ pub enum Route<'a> {
         /// ID of the guild.
         guild_id: u64,
         /// Maximum amount of members to get.
-        limit: Option<u64>,
+        limit: Option<u16>,
         /// ID of the scheduled event.
         scheduled_event_id: u64,
         /// Whether to return a member object.
@@ -594,7 +594,7 @@ pub enum Route<'a> {
         /// The maximum ID of guilds to get.
         before: Option<u64>,
         /// The maximum number of guilds to get.
-        limit: Option<u64>,
+        limit: Option<u16>,
     },
     /// Route information to get an original interaction response message.
     GetInteractionOriginal {
@@ -644,7 +644,7 @@ pub enum Route<'a> {
         /// The ID of the channel.
         channel_id: u64,
         /// The maximum number of messages to get.
-        limit: Option<u64>,
+        limit: Option<u16>,
     },
     /// Route information to get a list of sticker packs available to Nitro
     /// subscribers.
@@ -691,7 +691,7 @@ pub enum Route<'a> {
         /// The URI encoded custom or unicode emoji.
         emoji: &'a RequestReactionType<'a>,
         /// The maximum number of users to retrieve.
-        limit: Option<u64>,
+        limit: Option<u16>,
         /// The ID of the message.
         message_id: u64,
     },
@@ -814,7 +814,7 @@ pub enum Route<'a> {
         /// ID of the guild to search in.
         guild_id: u64,
         /// Upper limit of members to query for.
-        limit: Option<u64>,
+        limit: Option<u16>,
         /// Query to search by.
         query: &'a str,
     },
